Features/add netherlands march agenda (#6)
* WIP: Add the Dutch March event

Added the known data I got so far on the speaker, still have to update the topic of Pauls talk.

* Added info on the final speaker.

I think this agenda now represents for a big part the actual agenda. Feel free to edit the text you use on EventBrite as English is not my first language.